Refactor Business Component
Refactored business component to take in Business from props. Needed to do this in order to split the business component's use across business dashboard and business page
